requirements.txt
PyOpenGL
PyOpenGL_accelerate
glfw
numpy
Pillow
PyGLM